Store the lump's type as a field of the struct
The type is needed to be able to skip the entities lump without feeling
too hard-coded (which would be skipping it during reading the header).
Since the lumps array will be sorted directly, there won't be anything
keeping track of the original index if it's not stored in the struct.